Refac(server): Tests (#22)
* Refac(server): Better testing and WebSocket connection handling.
- Tests are no longer using replicated logic (which was obsolete, therefore testing nothing)
- WebSocket connections are now closed on client side upon Drop. This wasn't happening before since it was working under the assumption we'd either sig/term/kill always. Obviously that may not be the case, as is with the test where we attempt to drop connections and then submit() a proof request.
- In order to support the above, we now have a  wrapper type that implements oneshot communication, triggered on drop.

Current set of tests has its own implementation detached from the actual server impl. Meaning changes to the server are not directly mirrorred within the tests.

Here, we:

Eliminate that behaviour, make tests actually run by the server logic.
Current tests impl'd are focusing on the client side of sending packets with http multipart and receiving it via WebSockets, serialized and compressed.
Dockerfile and compose for spawning a server instance for tests.